# perf: improve startup time by doing some imports lazily