--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -33,7 +33,7 @@ jobs:
         python-version: 3.9
     - name: Install dependencies
       run: |
-        python -m pip install --disable-pip-version-check black isort
+        python -m pip install --disable-pip-version-check black==22.12.0 isort==5.11.4
     - name: Autoformatter
       run: |
         black --diff .