--- a/imaginairy/__init__.py
+++ b/imaginairy/__init__.py
@@ -1,19 +1,12 @@
-import os.path
+import os
 
 # tells pytorch to allow MPS usage (for Mac M1 compatibility)
 os.putenv("PYTORCH_ENABLE_MPS_FALLBACK", "1")
-import PIL.Image  # noqa
 
 from .api import imagine, imagine_image_files  # noqa
-from .enhancers.describe_image_blip import generate_caption  # noqa
 from .schema import (  # noqa
     ImaginePrompt,
     ImagineResult,
     LazyLoadingImage,
     WeightedPrompt,
 )
-from .version import __version__  # noqa
-
-# https://stackoverflow.com/questions/71738218/module-pil-has-not-attribute-resampling
-if not hasattr(PIL.Image, "Resampling"):  # Pillow<9.0
-    PIL.Image.Resampling = PIL.Image