--- a/imaginairy/bin/imagine
+++ b/imaginairy/bin/imagine
@@ -1,0 +1,6 @@
+#!/bin/env python
+# -*- coding: utf-8 -*-
+
+if __name__ == "__main__":
+    from imaginairy import cmds
+    cmds.imagine_cmd()