--- a/imaginairy/log_utils.py
+++ b/imaginairy/log_utils.py
@@ -4,12 +4,6 @@ import re
 import time
 import warnings
 
-import torch
-from pytorch_lightning import _logger as pytorch_logger
-from torchvision.transforms import ToPILImage
-from transformers.modeling_utils import logger as modeling_logger
-from transformers.utils.logging import _configure_library_root_logger
-
 _CURRENT_LOGGING_CONTEXT = None
 
 logger = logging.getLogger(__name__)
@@ -159,6 +153,9 @@ class ImageLoggingContext:
     def log_img(self, img, description):
         if not self.debug_img_callback:
             return
+        import torch
+        from torchvision.transforms import ToPILImage
+
         self.image_count += 1
         if isinstance(img, torch.Tensor):
             img = ToPILImage()(img.squeeze().cpu().detach())
@@ -200,6 +197,8 @@ def filesafe_text(t):
 
 
 def conditioning_to_img(conditioning):
+    from torchvision.transforms import ToPILImage
+
     return ToPILImage()(conditioning)
 
 
@@ -252,6 +251,9 @@ def configure_logging(level="INFO"):
 
 
 def disable_transformers_custom_logging():
+    from transformers.modeling_utils import logger as modeling_logger
+    from transformers.utils.logging import _configure_library_root_logger
+
     _configure_library_root_logger()
     _logger = modeling_logger.parent
     _logger.handlers = []
@@ -263,6 +265,8 @@ def disable_transformers_custom_logging():
 
 
 def disable_pytorch_lighting_custom_logging():
+    from pytorch_lightning import _logger as pytorch_logger
+
     try:
         from pytorch_lightning.utilities.seed import log  # noqa