--- a/tox.ini
+++ b/tox.ini
@@ -11,7 +11,7 @@ format = pylint
 skip = */.tox/*,*/.env/*,build/*,*/downloads/*,other/*,prolly_delete/*,downloads/*,imaginairy/vendored/*,testing_support/vastai_cli_official.py,.eggs/*
 linters = pylint,pycodestyle,pyflakes,mypy
 ignore =
-    Z999,C0103,C0301,C0302,C0114,C0115,C0116,
+    Z999,C0103,C0301,C0302,C0114,C0115,C0116,C0415,
     Z999,D100,D101,D102,D103,D105,D106,D107,D200,D202,D203,D205,D212,D400,D401,D406,D407,D413,D415,D417,
     Z999,E203,E501,E1101,E1131,E1135,E1136,
     Z999,R0901,R0902,R0903,R0904,R0193,R0912,R0913,R0914,R0915,R1702,